vhost: change vsession->id to uint64_t
This id is assigned from a global increment-only counter
whose type is uint64_t. I believe the original intent
was to use uint64_t for vsession->id as well, let's fix
it now to avoid wrapping errors.

Acked-by: Vitaliy Mysak <vitaliy.mysak@intel.com>

+++ lib/vhost/vhost_internal.h
@@ -126,7 +126,7 @@ struct spdk_vhost_session {
 	int vid;
 
 	/* Unique session ID. */
-	unsigned id;
+	uint64_t id;
 
 	struct vhost_poll_group *poll_group;
